Work on First Stage continued
Edited the straight created to now have a new sidewalk texture.

Created 90 and 180 degree turn prefabs

Started work on a sidestep section

Continued building the first stage

Created a level gate which will load the stage from the hub-world

Linked the first hub-world to the new game option in the Main game menu.

Added a motion-blur effect in post-processing.