Final Commit
Please note a lot of this work was done during April and may but was only commit now due to storage limitations on GitHub servers. This meant that in order to upload changes, all the "library" folders had to be deleted, and then recompiled by Unity if the project was opened, so therefore it became a very time consuming process.

The shopping system has been introduced, allowing players to buy and equip costumes.

Re arranged controller bindings for xbox controllers.

Fixed collision issues in hub world and action stage.

XP rewards implemented based on the players score and rank.

Created the Final build of the project